requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
flask>=2.0.0
werkzeug>=2.0.0
gunicorn>=20.1.0 